refactor Dashboard imports
